requirements.txt
pyTelegramBotAPI==4.14.0
feedparser==6.0.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
urllib3==2.0.7
backoff==2.2.1

pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

black==23.12.1
flake8==6.1.0
isort==5.13.2
pylint==3.0.3
mypy==1.8.0

bandit==1.7.6
safety==3.2.11